requirements.txt
flask
python-dotenv
paho-mqtt
Pillow
python-telegram-bot